+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed effect chain routing updates on state changes by expanding the nodes helper with iblokz-audio node-level routing logic
+
 ## [1.0.1] - 2026-01-17
 
 ### Fixed